test: add metadata-uri + uri-storage tests
