+++ examples/erc1155-metadata-uri/tests/abi/mod.rs
@@ -4,34 +4,13 @@ use alloy::sol;
 sol!(
     #[sol(rpc)]
     contract Erc1155 {
-        function balanceOf(address account, uint256 id) external view returns (uint256 balance);
         #[derive(Debug)]
-        function balanceOfBatch(address[] accounts, uint256[] ids) external view returns (uint256[] memory balances);
-        function isApprovedForAll(address account, address operator) external view returns (bool approved);
-        function setApprovalForAll(address operator, bool approved) external;
-        function safeTransferFrom(address from, address to, uint256 id, uint256 value, bytes memory data) external;
-        function safeBatchTransferFrom(address from, address to, uint256[] memory ids, uint256[] memory values, bytes memory data) external;
-        function mint(address to, uint256 id, uint256 amount, bytes memory data) external;
-        function burn(address account, uint256 id, uint256 value) external;
         function uri(uint256 id) external view returns (string memory uri);
         function setTokenURI(uint256 tokenId, string memory tokenURI) external;
 
         function supportsInterface(bytes4 interfaceId) external view returns (bool);
 
-        error ERC1155InvalidArrayLength(uint256 idsLength, uint256 valuesLength);
-        error ERC1155InvalidOperator(address operator);
-        error ERC1155InvalidSender(address sender);
-        error ERC1155InvalidReceiver(address receiver);
-        error ERC1155MissingApprovalForAll(address operator, address owner);
-        error ERC1155InsufficientBalance(address sender, uint256 balance, uint256 needed, uint256 tokenId);
-
-        #[derive(Debug, PartialEq)]
-        event TransferSingle(address indexed operator, address indexed from, address indexed to, uint256 id, uint256 value);
-        #[derive(Debug, PartialEq)]
-        event TransferBatch(address indexed operator, address indexed from, address indexed to, uint256[] ids, uint256[] values);
-        #[derive(Debug, PartialEq)]
-        event ApprovalForAll(address indexed account, address indexed operator, bool approved);
         #[derive(Debug, PartialEq)]
         event URI(string value, uint256 indexed id);
-}
+    }
 );
